rag-backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
chromadb
langchain-text-splitters
groq
pymupdf
python-pptx
tqdm
requests
